rm openai_utils, moves it to "utils'
